+++ CHANGELOG.md
@@ -6,8 +6,9 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 ## [3.3.0] - 2026-03-31
 
 * Awesome new change indicators per section, and diff view per section (thanks Leo!)
+* Insert section gesture! Keep your mouse over the middle of two sections for 1sec and you'll see it.
 * Finished the implementation of the various axis and other settings for Sankey and Heatmap chart types
-* Bugs: 
+* Bugs:
     * Added new chart types and forgot to update the Copilot tools to include them. D'oh.
     * Fixed spacing issues with the title and a couple of the chart types that overlapped their content.